[core/refactor] Split NonEmptyReducible from Reducible
